Completed EnemySpawner
Can set the object to spawn, whether to use the Spawners location as the origin to modify spawn origin in editor, whether to spawn repeatedly, a custom origin, a xyz random range to spawn within, the number of a given class to spawn, and how long to wait before spawning again if spawning repeatedly

Note: Currently the value for range to randomize for a small given area is 300. Range in general should be in the hundreds to see results / non-overlapping game objects.